+++ fabfile/docs.py
@@ -5,9 +5,6 @@ from fabric.contrib.project import rsync_project
 from fabric.version import get_version
 
 
-docs_host = 'jforcier@fabfile.org'
-
-
 @task(default=True)
 def build(clean='no', browse_='no'):
     """
@@ -29,14 +26,3 @@ def browse():
     Open the current dev docs in a browser tab.
     """
     local("open docs/_build/html/index.html")
-
-
-@task
-@hosts(docs_host)
-def push():
-    """
-    Build docs and zip for upload to RTD
-    """
-    build(clean='yes')
-    v = get_version('short')
-    local("cd docs/_build/html && zip -r ../%s.zip ." % v)