+++ tests/utils.py
@@ -16,6 +16,7 @@ import tempfile
 
 from fudge import Fake, patched_context, clear_expectations
 from nose.tools import raises
+from nose import SkipTest
 
 from fabric.context_managers import settings
 from fabric.state import env, output
@@ -38,6 +39,13 @@ class FabricTest(object):
         # Deepcopy doesn't work well on AliasDicts; but they're only one layer
         # deep anyways, so...
         self.previous_output = output.items()
+        # Allow hooks from subclasses here for setting env vars (so they get
+        # purged correctly in teardown())
+        self.env_setup()
+        # Temporary local file dir
+        self.tmpdir = tempfile.mkdtemp()
+
+    def env_setup(self):
         # Set up default networking for test server
         env.disable_known_hosts = True
         env.update(to_dict('%s@%s:%s' % (USER, HOST, PORT)))
@@ -45,10 +53,9 @@ class FabricTest(object):
         # Command response mocking is easier without having to account for
         # shell wrapping everywhere.
         env.use_shell = False
-        # Temporary local file dir
-        self.tmpdir = tempfile.mkdtemp()
 
     def teardown(self):
+        env.clear() # In case tests set env vars that didn't exist previously
         env.update(self.previous_env)
         output.update(self.previous_output)
         shutil.rmtree(self.tmpdir)
@@ -261,8 +268,10 @@ def patched_env(updates):
     return wrapper
 
 
-def fabfile(name):
-    return os.path.join(os.path.dirname(__file__), 'support', name)
+def support(path):
+    return os.path.join(os.path.dirname(__file__), 'support', path)
+
+fabfile = support
 
 
 @contextmanager